+++ src/litgen/internal/adapted_types/adapted_class.py
@@ -1045,9 +1045,6 @@ class AdaptedClass(AdaptedElement):
         if not self._virt_shall_override():
             return
 
-        virtual_methods = self._virt_method_list_including_inherited()
-        nb_virtual_methods = len(virtual_methods)
-
         if self.options.bind_library == litgen.BindLibraryType.pybind11:
             trampoline_class_template = code_utils.unindent_code(
                 """
@@ -1069,7 +1066,7 @@ class AdaptedClass(AdaptedElement):
                 class {trampoline_class_name} : public {class_name}
                 {
                 public:
-                    NB_TRAMPOLINE({class_name}, {nb_virtual_methods});
+                    NB_TRAMPOLINE({class_name});
 
                 {trampoline_list}
                 };
@@ -1089,7 +1086,6 @@ class AdaptedClass(AdaptedElement):
 
         replacements = munch.Munch()
         replacements.trampoline_class_name = self.cpp_element().class_name + "_trampoline"
-        replacements.nb_virtual_methods = str(nb_virtual_methods)
         replacements.class_name = self.cpp_element().class_name
         replacements.trampoline_list = code_utils.indent_code(
             "\n".join(trampoline_lines), indent_str=self.options._indent_cpp_spaces()